Add files via upload
so i had back and forth with classmates on how question 1 part 2 is done, and i got a simulation to get an answer. however the answers don't match the simulation, and i couldn't get it to work. this is so i can ask smay or ashley later on what i should have done to make it work, or if i did it wrong.